Refactoring MDM code
Jira-Id: SECO-3163
