+++ modules/sc-mesh-secure-deployment/src/nats/src/comms_controller.py
@@ -1,0 +1,52 @@
+import logging
+from . import comms_command
+from . import comms_settings
+from . import comms_status
+from . import comms_mesh_telemetry
+
+
+class CommsController:  # pylint: disable=too-few-public-methods
+    """
+    Mesh network
+    """
+
+    def __init__(self, interval: int = 1000):
+        self.interval = interval
+        self.debug_mode_enabled = False
+
+        # base logger for comms and which is used by all other modules
+        self.main_logger = logging.getLogger("comms")
+        self.main_logger.setLevel(logging.DEBUG)
+        log_formatter = logging.Formatter(
+            fmt="%(asctime)s :: %(name)-18s :: %(levelname)-8s :: %(message)s"
+        )
+        console_handler = logging.StreamHandler()
+        console_handler.setFormatter(log_formatter)
+        self.main_logger.addHandler(console_handler)
+
+        self.c_status = []
+        # TODO how many radios?
+        for i in range(0, 3):
+            self.c_status.append(
+                comms_status.CommsStatus(
+                    self.main_logger.getChild(f"status {str(i)}"), i
+                )
+            )
+
+        self.settings = comms_settings.CommsSettings(
+            self.c_status, self.main_logger.getChild("settings")
+        )
+
+        for cstat in self.c_status:
+            if cstat.index < len(self.settings.mesh_vif):
+                cstat.wifi_interface = self.settings.mesh_vif[cstat.index]
+
+        self.command = comms_command.Command(
+            self.c_status, self.main_logger.getChild("command")
+        )
+        self.telemetry = comms_mesh_telemetry.MeshTelemetry(
+            self.interval, self.main_logger.getChild("telemetry")
+        )
+
+        # logger for this module and derived from main logger
+        self.logger = self.main_logger.getChild("controller")
